Refactor TFM updating code
Move the code for updating TFMs into a helper and re-use from the two upgraders.
